feat: move inner class to upper level (#10)
