+++ src/test/java/com/shyoon/wms/location/feature/AssignLocationLPNTest.java
@@ -1,21 +1,22 @@
 package com.shyoon.wms.location.feature;
 
-import com.shyoon.wms.inbound.domain.LPN;
 import com.shyoon.wms.inbound.domain.LPNRepository;
-import com.shyoon.wms.location.domain.Location;
 import com.shyoon.wms.location.domain.LocationRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.util.Assert;
 
 class AssignLocationLPNTest {
 
     private AssignLocationLPN assignLocationLPN;
+    private LocationRepository locationRepository;
+    private LPNRepository lpnRepository;
 
     @BeforeEach
     void setUp() {
-        assignLocationLPN = new AssignLocationLPN();
+        assignLocationLPN = new AssignLocationLPN(
+                locationRepository,
+                lpnRepository);
     }
 
     @Test
@@ -32,28 +33,4 @@ class AssignLocationLPNTest {
 
     }
 
-    private class AssignLocationLPN {
-
-        private LocationRepository locationRepository;
-        private LPNRepository lpnRepository;
-
-        public void request(final Request request) {
-            final Location location = locationRepository.getByLocationBarcode(request.locationBarcode);
-            final LPN lpn = lpnRepository.getByLPNBarcode(request.lpnBarcode);
-
-            location.assignLPN(lpn);
-        }
-
-
-
-        public record Request(
-                String locationBarcode,
-                String lpnBarcode) {
-
-            public Request {
-                Assert.hasText(locationBarcode, "로케이션 바코드는 필수입니다.");
-                Assert.hasText(lpnBarcode, "LPN 바코드는 필수입니다.");
-            }
-        }
-    }
 }
